+++ src/onegov.election_day/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Fixes district/entity not shown in election compounds PDF.
+  [msom]
+
 3.4.2 (2018-04-10)
 ~~~~~~~~~~~~~~~~~~~